Extract interpreter
